AI_Model_CSR_Denial_Knowledge_Bot/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
sentence-transformers==2.6.1
faiss-cpu==1.8.0
pandas==2.2.1
numpy==1.26.4
flask==3.0.0
flask-cors==4.0.0
requests==2.32.3
gunicorn==21.2.0
